Fixed: The navigationView and rightDrawerView was overlapping on statusBar.
