+++ core/src/main/java/org/kiwix/kiwixmobile/core/base/BaseFragment.kt
@@ -21,16 +21,14 @@ package org.kiwix.kiwixmobile.core.base
 import android.content.Context
 import android.os.Bundle
 import android.view.View
-import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import org.kiwix.kiwixmobile.core.R
 import org.kiwix.kiwixmobile.core.extensions.enableEdgeToEdgeMode
 import org.kiwix.kiwixmobile.core.extensions.getToolbarNavigationIcon
 import org.kiwix.kiwixmobile.core.extensions.setToolTipWithContentDescription
+import org.kiwix.kiwixmobile.core.extensions.setTopMarginToViewForEdgeToEdgeMode
 
 /**
  * All fragments should inherit from this fragment.
@@ -52,15 +50,7 @@ abstract class BaseFragment : Fragment() {
     super.onViewCreated(view, savedInstanceState)
     enableEdgeToEdgeMode()
     setupToolbar()
-    this.view?.let {
-      ViewCompat.setOnApplyWindowInsetsListener(it) { view, windowInsets ->
-        val insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars())
-        val layoutParams = view.layoutParams as? ViewGroup.MarginLayoutParams
-        layoutParams?.topMargin = insets.top
-        view.layoutParams = layoutParams
-        WindowInsetsCompat.CONSUMED
-      }
-    }
+    this.view.setTopMarginToViewForEdgeToEdgeMode()
   }
 
   // Setup toolbar to handle common back pressed event
